Extracted Master actions into separate partial Hub
